requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.95.0
anyio==4.13.0
apache-tvm-ffi==0.1.10
astor==0.8.1
attrs==26.1.0
bitsandbytes==0.49.2
blake3==1.0.8
cachetools==7.0.5
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
compressed-tensors==0.14.0.1
cramjam==2.11.0
cryptography==46.0.7
cuda-bindings==13.0.3
cuda-pathfinder==1.5.3
cuda-python==13.0.3
datasets==4.8.4
depyf==0.20.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.18.0
einops==0.8.2
email-validator==2.3.0
fastapi==0.135.3
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.0
fastar==0.11.0
fastparquet==2026.3.0
filelock==3.28.0
flashinfer-cubin==0.6.6
flashinfer-python==0.6.6
frozenlist==1.8.0
fsspec==2026.3.0
gguf==0.18.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.10.2
idna==3.11
ijson==3.5.0
importlib-metadata==8.7.1
interegular==0.3.3
jinja2==3.1.6
jiter==0.14.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.2.2
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
mistral-common==1.11.0
model-hosting-container-standards==0.1.14
mpmath==1.3.0
msgspec==0.21.1
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
ninja==1.13.0
numba==0.61.2
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.15.1.9
nvidia-cudnn-frontend==1.18.0
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-cutlass-dsl==4.4.2
nvidia-cutlass-dsl-libs-base==4.4.2
nvidia-ml-py==13.595.45
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
openai==2.31.0
openai-harmony==0.0.8
opencv-python-headless==4.13.0.92
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-exporter-otlp-proto-http==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
opentelemetry-semantic-conventions-ai==0.5.1
outlines-core==0.2.11
packaging==26.1
pandas==3.0.1
partial-json-parser==0.2.1.1.post7
peft==0.18.1
pillow==12.2.0
prometheus-client==0.25.0
prometheus-fastapi-instrumentator==7.1.0
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==23.0.1
pybase64==1.4.3
pycountry==26.2.16
pycparser==3.0
pydantic==2.13.1
pydantic-core==2.46.1
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pygments==2.20.0
pyjwt==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.1.0
python-multipart==0.0.26
pyyaml==6.0.3
pyzmq==27.1.0
quack-kernels==0.3.10
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
sentry-sdk==2.58.0
setproctitle==1.3.7
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.3.4
starlette==0.52.1
supervisor==4.3.0
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0+cu130
torch-c-dlpack-ext==0.1.5
torchaudio==2.10.0+cu130
torchvision==0.25.0+cu130
tqdm==4.67.3
transformers==5.5.4
triton==3.6.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
vllm==0.19.0
llama-cpp-python>=0.3
watchfiles==1.1.1
websockets==16.0
xgrammar==0.1.33
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
